+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updated Polars to 0.27.1
 - Added `iter_rows` method to `DataFrame`
 - Added `named` option to `row` and `rows` methods
+- Made `Series` and `GroupBy` enumerable
 - Replaced `include_bounds` option with `closed` for `is_between` method
 
 ## 0.2.5 (2023-02-01)